CONSOLE-5015: Convert name-value-editor to TypeScript functional components
Replace class components with FC<Props> using TypeScript interfaces,
react-dnd hooks (useDrag/useDrop), and useTranslation hook. Also fixes
knative plugin passing non-existent addLabel prop instead of addString.
